+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v1.4.0
 * Added a validation check for `id` properties in objects within arrays in data schemas, this will now be reported as an error if present.
+* Added validation for duplicate transition names in data schemas, this will now be reported as an error if present.
 
 ### v1.3.0
 * Added a command to synchronize Dispatchers `exh dispatchers sync --file=<path>` the file argument must point to a JSON file containing an array of Dispatchers to be synchronized.